requirements.txt
pyserial
tk
numpy
http://bit.ly/csc161graphics